cloud_functions/requirements.txt
flask >= 2.2.2, <2.3.0 
numpy >= 1.23.3, <1.24.0
scikit-learn >=1.1.2, <1.2.0
google-cloud-storage >=2.5.0, <2.6.0
